# week3/api/openapi.yaml
openapi: 3.0.3

info:
  title: Cloud3 Managed PostgreSQL API
  version: 0.1.0
  description: |
    REST API for provisioning and managing CloudNativePG PostgreSQL clusters
    on Kubernetes.

servers:
  - url: http://localhost:8080
    description: Local development or kubectl port-forward
  - url: http://cloud4-api.postgres-demo.svc.cluster.local
    description: Kubernetes cluster-internal service


tags:
  - name: Health
  - name: Authentication
  - name: Instances
  - name: Connection
  - name: Credentials

# all endpoints required token by default 
security:
  - bearerAuth: []

# endpoints
paths:
  # login endpoints
  /api/v1/auth/login:
    post:
      tags: [Authentication]
      summary: Log in
      description: Authenticate a demo user and return a JWT access token.
      operationId: login
      security: []  # means no token required for login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /api/v1/health:
    get:
      tags: [Health]
      summary: Check API health
      operationId: getHealth
      security: []
      responses:
        "200":
          description: API is running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /api/v1/instances:
    get:
      tags: [Instances]
      summary: List managed PostgreSQL instances
      operationId: listInstances
      responses:
        "200":
          description: Instances were retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DBInstance"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    post:
      tags: [Instances]
      summary: Create a managed PostgreSQL instance
      operationId: createInstance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInstanceRequest"
            example:
              name: pet-health-database
              instances: 1
      responses:
        "201":
          description: Cluster custom resource was created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DBInstance"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalServerError"


  /api/v1/instances/{id}:
    parameters:
      - $ref: "#/components/parameters/InstanceID"


    get:
      tags: [Instances]
      summary: Get one managed PostgreSQL instance
      operationId: getInstance
      responses:
        "200":
          description: Instance was retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DBInstance"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"


    # put:
    #   tags: [Instances]
    #   summary: Update a managed PostgreSQL instance
    #   operationId: updateInstance
    #   requestBody:
    #     required: true
    #     content:
    #       application/json:
    #         schema:
    #           $ref: "#/components/schemas/UpdateInstanceRequest"
    #         example:
    #           name: pet-health-database-production
    #           instances: 2
    #   responses:
    #     "200":
    #       description: Instance was updated successfully
    #       content:
    #         application/json:
    #           schema:
    #             $ref: "#/components/schemas/DBInstance"
    #     "400":
    #       $ref: "#/components/responses/BadRequest"
    #     "404":
    #       $ref: "#/components/responses/NotFound"
    #     "415":
    #       $ref: "#/components/responses/UnsupportedMediaType"
    #     "500":
    #       $ref: "#/components/responses/InternalServerError"
    patch:
      tags: [Instances]
      summary: Partially update an instance
      operationId: patchInstance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchInstanceRequest"
      responses:
        "200":
          description: Instance update accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DBInstance"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      tags: [Instances]
      summary: Delete a managed PostgreSQL instance
      operationId: deleteInstance
      responses:
        "202":
          description: Instance deletion request accepted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"




  /api/v1/instances/{id}/connection:
    parameters:
      - $ref: "#/components/parameters/InstanceID"

    get:
      tags: [Connection]
      summary: Get PostgreSQL connection information
      description: |
        Reads the CloudNativePG read-write Service and application Secret.
        The returned host is intended for clients running inside the
        Kubernetes cluster.
      operationId: getInstanceConnection
      responses:
        "200":
          description: Connection information was retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionInfo"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: The operator has not created the Service or Secret yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: connection information is not ready
        "500":
          $ref: "#/components/responses/InternalServerError"








# schema
components:
  parameters:
    InstanceID:
      name: id
      in: path
      required: true
      description: Kubernetes metadata.name of the CloudNativePG Cluster
      schema:
        type: string
        example: db-k9zhw


  securitySchemes: 
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT returned by POST /api/v1/auth/login




  schemas:
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 1
          example: admin
        password:
          Type: string
          format: password
          minLength: 1
          example: admin-pass

    LoginResponse:
      type: object
      required:
        - accessToken
        - tokenType
        - expiresIn
        - role
      properties:
        accessToken:
          type: string
          description: Signed JWT access token
        tokenType:
          type: string
          enum:
            - Bearer
          example: Bearer
        expiresIn:
          type: integer
          description: Token lifetime in seconds
          example: 3600
        role:
          $ref: "#/components/schemas/UserRole"

    UserRole:
      type: string
      enum:
        - admin
        - viewer
      example: viewer

    HealthResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          example: ok
   
    DBInstance:
      type: object
      required: [id, name, instances, storage, cpu, status, createdAt]
      properties:
        id:
          type: string
          example: db-k9zhw
        name:
          type: string
          example: pet-health-database
        instances:
          type: integer
          minimum: 1
          example: 2
        storage:
          type: string
          description: Storage size allocated to each PostgreSQL instance
          example: 5Gi
        cpu:
          type: string
          description: CPU request for each PostgreSQL Pod
          example: 500m
        status:
          type: string
          example: Cluster in healthy state
        createdAt:
          type: string
          format: date-time
          example: "2026-08-05T11:36:16Z"

    CreateInstanceRequest:
      type: object
      required: [name, instances]
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          example: pet-health-database
        instances:
          type: integer
          minimum: 1
          example: 1
        storage:
          type: string
          minLength: 1
          description: Initial storage size for each PostgreSQL instance
          default: 1Gi 
          example: 5Gi
        cpu:
          type: string
          minLength: 1
          description: Initial CPU request for each PostgreSQL Pod
          default: 250m
          example: 500m

    # UpdateInstanceRequest:
    #   type: object
    #   required: [name, instances]
    #   additionalProperties: false
    #   properties:
    #     name:
    #       type: string
    #       minLength: 1
    #       example: pet-health-database-production
    #     instances:
    #       type: integer
    #       minimum: 1
    #       example: 2

    PatchInstanceRequest:
      type: object
      minProperties: 1 #at least change one property
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          example: production-database
        instances:
          type: integer
          minimum: 1
          example: 3
        storage:
          type: string
          minLength: 1
          description: Desired storage size. Storage can only be increased.
          example: 5Gi
        cpu:
          type: string
          minLength: 1
          description: CPU request for each PostgreSQL Pod, not API Pods.
          example: 500m


    ConnectionInfo:
      type: object
      required: [host, port, database, username, password]
      properties:
        host:
          type: string
          description: Cluster-internal read-write Service host
          example: db-k9zhw-rw
        port:
          type: string
        database:
          type: string
        username:
          type: string
        password:
          type: string
          format: password
          description: Generated database password; do not log or commit it
        uri:
          type: string
          format: uri
          description: PostgreSQL connection URI
          example: postgresql://app:REDACTED@db-k9zhw-rw.postgres-demo:5432/app


    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: instance not found



# status code
  responses:
    BadRequest:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: invalid request body

    UnsupportedMediaType:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Content-Type is not application/json

    NotFound:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: instance not found

    InternalServerError:
      description: An internal operation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: internal server error

    Unauthorized:
      description: Authentication is required or the access token is invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: unauthorized

    Forbidden:
      description: The authenticated user does not have permission for this operation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: forbidden